requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bleak==1.0.1
camel-ai==0.2.70
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
docstring-parser==0.15
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mcp==1.12.2
openai==1.97.1
pillow==10.4.0
psutil==5.9.8
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-CoreBluetooth==11.1
pyobjc-framework-libdispatch==11.1
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.26.0
setuptools==69.5.1
sniffio==1.3.1
sse-starlette==3.0.0
starlette==0.47.2
tiktoken==0.7.0
toio.py==1.1.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
